EDSC-4543: Fix most of the tests skip some as a quick test against GH
EDSC-4543: Fix most of the tests skip some as a quick test against GH

EDSC-4543: Split out some playwright tests
